+++ test/integration/helpers/serverspec/service_examples.rb
@@ -3,6 +3,7 @@ require_relative 'spec_helper'
 shared_examples_for 'elasticsearch service' do |service_name = 'elasticsearch', args = {}|
   content_match = args[:content] || 'elasticsearch'
 
+  # we should move to inspec here ASAP, as this doesn't pass due to serverspec
   describe service(service_name) do
     it { should be_enabled }
     it { should be_running }